% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{convert_natcode_to_geocode}
\alias{convert_natcode_to_geocode}
\title{Convert Natcode to Geocode}
\usage{
convert_natcode_to_geocode(natcode = NULL, geocode = NULL,
  exponential_notation = FALSE)
}
\arguments{
\item{natcode}{(int) geographical code from INE}

\item{geocode}{(string) geographical code from Eurostat}

\item{exponential_notation}{(boolean) to show or not exponential notation. \code{e.g. e+10}}
}
\description{
This function allows converting from natcode to geocode and vice-versa.
If all params are null, you will get the complete table used for codes conversion.
}
\examples{
convert_natcode_to_geocode()
convert_natcode_to_geocode(natcode = 34050000000)
convert_natcode_to_geocode(geocode = "ES70")
}
