% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{draw_serie}
\alias{draw_serie}
\title{Draw serie}
\usage{
draw_serie(serie, geographical_granularity, nult = 0)
}
\arguments{
\item{serie}{(string) serie identificator}

\item{geographical_granularity}{(string) can be one of this: \code{comunidades_autonomas}, \code{provincias} or \code{municipios}}

\item{nult}{(int) last \code{n} serie data, if \code{nult = 0} this value will be auto-calculated}
}
\description{
This function allows representing series data into a map
}
\examples{
draw_serie("IPC251541", "provincias")
}
