% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_rds}
\alias{get_cache_rds}
\title{Get rds content (cache)}
\usage{
get_cache_rds(object, type = "SERIEOPERATION")
}
\arguments{
\item{object}{(string) an object of the "type" option}

\item{type}{(string) type of content do you want to read, \code{type = "/POLYGONS-"} by default}
}
\description{
This function returns the content of a RDS file in cache
}
\examples{
get_cache_rds("provincias", type = "POLYGONS")
get_cache_rds("comunidades_autonomas", type = "POLYGONS")
get_cache_rds("municipios", type = "POLYGONS")
get_cache_rds("natcodes", type = "DATATABLE")
get_cache_rds(4, type = "SERIEOPERATION")
}
