% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\name{get_data_metadataoperation}
\alias{get_data_metadataoperation}
\title{Get data metadata-operation}
\usage{
get_data_metadataoperation(code, query = NULL, p = 1, nult = 1, det = 0,
  tip = NA, ioe = FALSE, lang = "ES")
}
\arguments{
\item{code}{table identification}

\item{query}{string separated by \code{AND} with syntax \code{variable = value} using natural language}

\item{p}{periodicity}

\item{nult}{last \code{n} values}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = AM} to obtain the metadata (crossing variables-values) of the series and a friendly output.}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with latest \code{n} series data by crossing metadata from an id or code
}
\examples{
get_data_metadataoperation("IPC", query = "Provincias = Madrid AND Tipo de dato = Variacion mensual AND Grupos ECOICOP = NULL", nult = 1)
}
