% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_serie}
\alias{get_data_serie}
\title{Get data serie}
\usage{
get_data_serie(serie, date_start = NULL, date_end = NULL, nult = 0,
  det = 0, lang = "ES")
}
\arguments{
\item{serie}{serie identifier}

\item{date_start}{start date in format (string) \code{YYYY-MM-DD}}

\item{date_end}{end date in format (string) \code{YYYY-MM-DD}}

\item{nult}{last \code{n} values}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with data of a serie from an id or/and from a date or date range
}
\examples{
get_data_serie("IPC206449", nult = 1)
get_data_serie("IPC206449", nult = 5)
get_data_serie("IPC206449", "2013-01-01", "2016-01-01")
get_data_serie("IPC206449", "2010-01-01")
}
