% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_table}
\alias{get_data_table}
\title{Get data table}
\usage{
get_data_table(id, nult = 0, det = 0, tip = NULL, lang = "ES")
}
\arguments{
\item{id}{table identifier}

\item{nult}{last \code{n} values}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = AM} to obtain the metadata (crossing variables-values) of the series and a friendly output.}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with latest \code{n} data of a table from an id or code
}
\examples{
get_data_table(22350, 4)
}
