% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_frequency}
\alias{get_frequency}
\title{Get frequency}
\usage{
get_frequency(periodicity, data = NULL)
}
\arguments{
\item{periodicity}{string with a periodicity}

\item{data}{result of \code{get_data_serie()$Data} call}
}
\description{
This function return the periodicity (number) of a time serie
}
\examples{
get_frequency("Mensual")
get_frequency("Trimestral")
get_frequency("Anual")
}
