% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_geographical}
\alias{get_geographical}
\title{Get geographical}
\usage{
get_geographical(
  code = NULL,
  resource = "variable",
  all = FALSE,
  verbose = FALSE,
  natcode = NULL,
  geocode = NULL,
  exponential_notation = FALSE,
  help = FALSE
)
}
\arguments{
\item{code}{(string) serie id}

\item{resource}{(string) resource to access, by default \code{resource = "variable"} to get serie metadata.
Possible values are \code{variable, natcode or natcode_to_geocode}}

\item{all}{(boolean) if \code{all = TRUE} you will get all natcodes}

\item{verbose}{(boolean) show more information during the process}

\item{natcode}{(int) geographical code from INE}

\item{geocode}{(string) geographical code from Eurostat}

\item{exponential_notation}{(boolean) to show or not exponential notation. \code{e.g. e+10}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}
}
\description{
This function returns geographical information
}
\examples{
get_geographical("IPC251522")
get_geographical(resource = "variable", help = TRUE)
get_geographical("IPC251522", resource = "natcode")
get_geographical(natcode = 34050000000, resource = "natcode_to_geocode")
}
