% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_geographical_variable}
\alias{get_geographical_variable}
\title{Get geographical variable}
\usage{
get_geographical_variable(serie)
}
\arguments{
\item{serie}{(string) serie id}
}
\description{
This function returns the geographical variable of a serie
}
\examples{
get_geographical_variable("IPC251522")
}
