% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_crossing.R
\name{get_metadata_crossing}
\alias{get_metadata_crossing}
\title{Get metadata crossing}
\usage{
get_metadata_crossing(code = NULL, resource = "series", help = FALSE,
  query = NULL, p = 1, det = 0, tip = NULL, nlast = 1,
  lang = "ES")
}
\arguments{
\item{code}{(string) operation identificator}

\item{resource}{(string) resource to access, by default \code{resource = "metadata"} to get serie metadata.
Possible values are \code{series or data}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}

\item{query}{(string) string separated by \code{AND} with syntax \code{variable = value} using natural language}

\item{p}{(int) periodicity, \code{p = 1} by default}

\item{det}{(int) \code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{(string) \code{tip = M} to obtain the metadata (crossing variables-values) of the series}

\item{nlast}{last \code{n} values}

\item{lang}{(string) language used to obtain information}
}
\description{
This function returns data or metadata by metadata crossing
}
\examples{
get_metadata_crossing()
get_metadata_crossing(resource = "series", help = TRUE)
get_metadata_crossing("IPC", resource = "series", query = "Provincias = Madrid AND Tipo de dato = Variación mensual AND Grupos ECOICOP = NULL")
get_metadata_crossing("IPC", resource = "data", query = "Provincias = Madrid AND Tipo de dato = Variacion mensual AND Grupos ECOICOP = NULL", nlast = 5)
}
