% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operations}
\alias{get_operations}
\title{Get operations}
\usage{
get_operations(code = NULL, resource = "all", help = FALSE,
  ioe = FALSE, geographical_granularity = NULL,
  temporal_granularity = NULL, verbose = TRUE, lang = "ES")
}
\arguments{
\item{code}{(string) operation identificator}

\item{resource}{(string) resource to access, by default \code{resource = "metadata"} to get serie metadata.
Possible values are \code{all, metadata, variables_values or by_granularity}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}

\item{ioe}{(boolean) \code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{geographical_granularity}{(string) geographical granularity}

\item{temporal_granularity}{(string) temporal granularity}

\item{verbose}{(boolean) show more information during the process}

\item{lang}{(string) language used to obtain information}
}
\description{
This function returns information about operations
}
\examples{
get_operations()
get_operations(resource = "all")
get_operations(resource = "all", help = TRUE)
get_operations("IPC", resource = "metadata")
get_operations("IPC", resource = "variables_values")
get_operations(resource = "by_granularity", geographical_granularity = "PROV")
}
