% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operations_all}
\alias{get_operations_all}
\title{Get operations (all)}
\usage{
get_operations_all(lang = "ES")
}
\arguments{
\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all available statistical operations
}
\examples{
get_operations_all()
}
