% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_operations_by_granularity}
\alias{get_operations_by_granularity}
\title{Get operations by granularity}
\usage{
get_operations_by_granularity(geographical_granularity = NULL,
  temporal_granularity = NULL, verbose = TRUE)
}
\arguments{
\item{geographical_granularity}{(string) geographical granularity}

\item{temporal_granularity}{(string) temporal granularity}

\item{verbose}{(boolean) show more information during the process}
}
\description{
This function returns a list of all operations that have a temporal or geographic granularity specified by user
}
\examples{
get_operations_by_granularity(geographical_granularity = "PROV")
get_operations_by_granularity(temporal_granularity = "Anual")
}
