% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{get_publications_date}
\alias{get_publications_date}
\title{Get publications date}
\usage{
get_publications_date(code, det = 0, lang = "ES")
}
\arguments{
\item{code}{publication identification}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all publication date of a publication from an id or code
}
\examples{
get_publications_date(8)
}
