% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{get_publications_operation}
\alias{get_publications_operation}
\title{Get publications operation}
\usage{
get_publications_operation(code, ioe = FALSE, det = 0, lang = "ES")
}
\arguments{
\item{code}{operation identification}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with publications of an operation from an id or code
}
\details{
Numeric code \code{id}
Alphabetic code \code{IPC}
\code{IOE} code (Inventario de Operaciones Estadísticas)
}
\examples{
get_publications_operation(25)
get_publications_operation(30138, ioe = TRUE)
}
