% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_rds_content}
\alias{get_rds_content}
\title{Get rds content (cache)}
\usage{
get_rds_content(object, type = "/POLYGONS-")
}
\arguments{
\item{object}{(string) an object of the "type" option}

\item{type}{(string) type of content do you want to read, \code{type = "/POLYGONS-"} by default}
}
\description{
This function returns the content of a RDS file in cache
}
\examples{
get_rds_content("provincias")
get_rds_content("comunidades_autonomas")
get_rds_content("municipios")
get_rds_content("natcodes", type = "/DATATABLE-")
}
