% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_serie_nult}
\alias{get_serie_nult}
\title{Get serie last-n}
\usage{
get_serie_nult(serie)
}
\arguments{
\item{serie}{(string) serie identificator}
}
\description{
This function returns the auto-calculated last-n series
}
\examples{
get_serie_nult("IPC251541")
}
