% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_series}
\alias{get_series}
\title{Get series}
\usage{
get_series(serie, resource = "metadata", det = 0, tip = NULL,
  lang = "ES")
}
\arguments{
\item{serie}{(string) serie identificator}

\item{resource}{(string) resource to access, by default \code{resource = "metadata"} to get serie metadata.
Possible values are \code{metadata, operation, values, table, metadataoperation, data, by_granularity, by_common_parameters or nult}}

\item{det}{(int) \code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{(string) \code{tip = M} to obtain the metadata (crossing variables-values) of the series}

\item{lang}{(string) language used to obtain information}
}
\description{
This function returns a data frame with a serie from an id or code
}
\examples{
get_series("IPC206449")
get_series("IPC206449", det = 2, tip = "M")
}
