% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_series_by_classification}
\alias{get_series_by_classification}
\title{Get series by classification}
\usage{
get_series_by_classification(serie, classification = NULL)
}
\arguments{
\item{serie}{(string) serie identification}

\item{classification}{(string) serie classification}
}
\description{
This function returns a list of all series of an operation that are in the same classification of a specific serie
}
\examples{
get_series_by_classification("IPC251539", classification = "Base 2016")
}
