% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_series_by_common_parameters}
\alias{get_series_by_common_parameters}
\title{Get series by common parameters}
\usage{
get_series_by_common_parameters(serie, classification = NULL,
  verbose = TRUE)
}
\arguments{
\item{serie}{(string) serie identification}

\item{classification}{(string) serie classification, if \code{classification = NULL} this value will be auto-detected}

\item{verbose}{(boolean) to show more information about this process, \code{verbose = TRUE} by default}
}
\description{
This function returns a list of all series of an operation that have the same classification, name and geographical granularity of a specific serie
}
\examples{
get_series_by_common_parameters("IPC251539", classification = "Base 1992")
get_series_by_common_parameters("IPC251539")
get_series_by_common_parameters("DPOP37286")
get_series_by_common_parameters("UA42121")
}
