% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_series_by_granularity}
\alias{get_series_by_granularity}
\title{Get series by granularity}
\usage{
get_series_by_granularity(operation, geographical_granularity = NULL,
  temporal_granularity = NULL, verbose = TRUE)
}
\arguments{
\item{operation}{(string) operation identification}

\item{geographical_granularity}{(string) geographical granularity}

\item{temporal_granularity}{(string) temporal granularity}

\item{verbose}{(boolean) show more information during the process}
}
\description{
This function returns a list of all series of an operation that has a temporal or geographic granularity specified by the user
}
\examples{
get_series_by_granularity("IPC", geographical_granularity = "PROV")
get_series_by_granularity("IPC", temporal_granularity = "Mensual")
get_series_by_granularity("IPC", geographical_granularity = "PROV", temporal_granularity = "Mensual")
}
