% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_series_operation}
\alias{get_series_operation}
\title{Get series operation}
\usage{
get_series_operation(code, det = 0, tip = NA, pagination = FALSE,
  page = NA, page_start = NA, page_end = NA, ioe = FALSE,
  lang = "ES", cache = TRUE, benchmark = FALSE)
}
\arguments{
\item{code}{operation identification}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = M} to obtain the metadata (crossing variables-values) of the series.}

\item{pagination}{\code{TRUE} to obtain data page by page and \code{FALSE} by default.}

\item{page}{\code{page = 1} to obtain data of an specific page (to use this, \code{pagination = FALSE}).}

\item{page_start}{\code{page_start = 1} start page range to obtain data (to use this, \code{pagination = TRUE}).}

\item{page_end}{\code{page_end = 2} end page range to obtain data (to use this, \code{pagination = TRUE}).}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}

\item{cache}{used to load data from local cache instead API, \code{cache = FALSE} by default.}

\item{benchmark}{used to measure the performance of the system, \code{benchmark = FALSE} by default.}
}
\description{
This function returns a data frame with all series of an operation from an id or code
}
\details{
Numeric code \code{id}
Alphabetic code \code{IPC}
\code{IOE} code (Inventario de Operaciones Estadísticas)
}
\examples{
get_series_operation(25)
get_series_operation(25, 2, "M")
get_series_operation(30138, ioe = TRUE)
get_series_operation(30138, 2, "M", ioe = TRUE)
get_series_operation("IPC", det = 2, tip = "M")
get_series_operation(25, pagination = FALSE, page = 1, cache = FALSE)
get_series_operation(25, pagination = TRUE, page_start = 1, page_end = 2, cache = FALSE)
}
