% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_series_values}
\alias{get_series_values}
\title{Get series values}
\usage{
get_series_values(code, det = 0, lang = "ES")
}
\arguments{
\item{code}{serie identification}

\item{det}{\code{det = 1} to see the detail of the variable to which it belongs, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with the metadata that defines a series from an id or code
}
\examples{
get_series_values("IPC206449")
get_series_values("IPC206449", 1)
}
