% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas.R
\name{get_tables}
\alias{get_tables}
\title{Get tables}
\usage{
get_tables(code = NULL, resource = "operation", help = FALSE,
  grp = NULL, geo = 0, nlast = 0, det = 0, tip = NULL,
  ioe = FALSE, lang = "ES")
}
\arguments{
\item{code}{operation (string/int) or table (int) identificator}

\item{resource}{(string) resource to access, by default \code{resource = "operation"} to get tables of an operation.
Possible values are \code{operation, group, group_values or data}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}}

\item{grp}{(int) group identification}

\item{geo}{(int) use \code{geo = 1} to access only tables with geographic content, \code{geo = 0} by default}

\item{nlast}{last \code{n} values}

\item{det}{(int) use \code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{(string) use \code{tip = "A"} to view as friendly, specifically the view the field \code{ultima_modificacion} or use \code{tip = "AM"}
to obtain the metadata (crossing variables-values) of the series and a friendly output.}

\item{ioe}{(boolean) use \code{ioe = TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{(string) language used to obtain information}
}
\description{
This function returns data / metadata from tables
}
\examples{
get_tables("IPC")
get_tables(resource = "operation", help = TRUE)
get_tables(25, resource = "operation")
get_tables(22350, resource = "group")
get_tables(22350, grp = 81497, resource = "group_values")
get_tables(22350, nlast = 5, resource = "data")
}
