% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas.R
\name{get_tables_group}
\alias{get_tables_group}
\title{Get tables group}
\usage{
get_tables_group(id, lang = "ES")
}
\arguments{
\item{id}{table identification}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with a combination of variables and values that define a table from an id or code
}
\examples{
get_tables_group(22350)
}
