% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas.R
\name{get_tables_group_values}
\alias{get_tables_group_values}
\title{Get tables group values}
\usage{
get_tables_group_values(id, grp, det = 0, lang = "ES")
}
\arguments{
\item{id}{table identification}

\item{grp}{group identification}

\item{det}{\code{det = 1} to see the detail of the variable to which it belongs, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with a combination of variables and values that define a table from an id or code
}
\examples{
get_tables_group_values(22350, 81497)
get_tables_group_values(22350, 81497, 1)
}
