% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablas.R
\name{get_tables_operation}
\alias{get_tables_operation}
\title{Get tables operation}
\usage{
get_tables_operation(code, det = 0, geo = 0, tip = NA, ioe = FALSE,
  lang = "ES")
}
\arguments{
\item{code}{operation identification}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{geo}{\code{geo = 1} to access only tables with geographic content, \code{geo = 0} by default}

\item{tip}{\code{tip = A} to view as friendly, specifically the view the field \code{ultima_modificacion}}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all tables of an operation with geographic information from an id or code
}
\details{
Numeric code \code{id}
Alphabetic code \code{IPC}
\code{IOE} code (Inventario de Operaciones Estadísticas)
}
\examples{
get_tables_operation(25)
get_tables_operation(25, 2, 1, tip = "A")
get_tables_operation(30138, ioe = TRUE, 2, 1, tip = "A")
}
