% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{get_values}
\alias{get_values}
\title{Get values}
\usage{
get_values(
  code = NULL,
  resource = "all",
  operation = NULL,
  help = FALSE,
  det = 0,
  ioe = FALSE,
  lang = "ES"
)
}
\arguments{
\item{code}{(int) variable indentificator}

\item{resource}{(string) resource to access, by default \code{resource = "all"} to get serie metadata.
Possible values are \code{all or variable_operation}}

\item{operation}{(string) operation indentificator}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}

\item{det}{(int) \code{det = 1} to see the detail of the variable to which it belongs, \code{det = 0} by default}

\item{ioe}{(boolean) \code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{(string) language used to obtain information}
}
\description{
This function returns values from a variable or variable operation
}
\examples{
get_values(115)
get_values(resource = "all", help = TRUE)
get_values(762, operation = "IPC", resource = "variable_operation")
}
