% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valores.R
\name{get_values_all}
\alias{get_values_all}
\title{Get values (all)}
\usage{
get_values_all(id, det = 0, lang = "ES")
}
\arguments{
\item{id}{operation identification}

\item{det}{\code{det = 1} to see the detail of the variable to which it belongs, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all values from a variable
}
\examples{
get_values_all(115)
get_values_all(115, 1)
}
