% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valores.R
\name{get_values_variableoperation}
\alias{get_values_variableoperation}
\title{Get values from variable operation}
\usage{
get_values_variableoperation(id, op, det = 0, ioe = FALSE,
  lang = "ES")
}
\arguments{
\item{id}{variable identification}

\item{op}{operation identification}

\item{det}{\code{det = 1} to see the detail of the variable to which it belongs, \code{det = 0} by default}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all values from a variable to an operation
}
\details{
Numeric code \code{id}
Alphabetic code \code{IPC}
\code{IOE} code (Inventario de Operaciones Estadísticas)
}
\examples{
get_values_variableoperation(762, 25)
get_values_variableoperation(762, 25, 1)
get_values_variableoperation(762, 30138, ioe = TRUE)
get_values_variableoperation(762, 30138, ioe = TRUE, 1)
}
