% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_variables}
\alias{get_variables}
\title{Get variables}
\usage{
get_variables(code = NULL, resource = "all", help = FALSE,
  ioe = FALSE, lang = "ES")
}
\arguments{
\item{code}{operation identifier}

\item{resource}{(string) resource to access, by default \code{resource = "metadata"} to get serie metadata.
Possible values are \code{all or operation}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}

\item{ioe}{(boolean) \code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns all or operations variables
}
\examples{
get_variables()
get_variables(resource = "all", help = TRUE)
get_variables("IPC", resource = "operation")
}
