% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_variables_all}
\alias{get_variables_all}
\title{Get variables (all)}
\usage{
get_variables_all(lang = "ES")
}
\arguments{
\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all system variables
}
\examples{
get_variables_all()
}
