% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{highcharts_series}
\alias{highcharts_series}
\title{Highcharts series}
\usage{
highcharts_series(code, date_start = NA, date_end = NA, nult = 0,
  det = 0, lang = "ES")
}
\arguments{
\item{code}{identification code of a serie (e.g. "IPC206449")}

\item{date_start}{start date in format (string) \code{YYYY-MM-DD}}

\item{date_end}{end date in format (string) \code{YYYY-MM-DD}}

\item{nult}{last \code{n} values}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function draws a highchart with data of a series from an id and/or from a date or date range
}
\examples{
highcharts_series("IPC206449", nult = 1) # Get the latest data of a series
highcharts_series("IPC206449", nult = 5) # Get the \\code{n} last data of a series
highcharts_series("IPC206449", "2013-01-01", "2016-01-01") # Get data of a series between two dates
highcharts_series("IPC206449", "2010-01-01") # Get data from a series from a date
}
