% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{represent_series}
\alias{represent_series}
\title{Represent series}
\usage{
represent_series(code = NULL, resource = "maps", help = FALSE,
  nlast = 0, date_start = NA, date_end = NA, classification = NULL,
  map_scale = 60, verbose = FALSE, benchmark = FALSE, det = 0,
  type = NA, lang = "ES")
}
\arguments{
\item{code}{(string) serie identificator}

\item{resource}{(string) resource to access, by default \code{resource = "all"} to get serie metadata.
Possible values are \code{maps, plot or highcharts}}

\item{help}{(boolean) type any value for \code{resource} param and type \code{help = TRUE} to see params available for this \code{resource}.}

\item{nlast}{(int) last \code{n} serie data, if \code{nult = 0} this value will be auto-calculated}

\item{date_start}{(date) start date in format (string) \code{YYYY-MM-DD}}

\item{date_end}{(date) end date in format (string) \code{YYYY-MM-DD}}

\item{classification}{(string) serie classification, if \code{classification = NULL} this value will be auto-detected}

\item{map_scale}{(int) refers to the relationship or ratio between distance on a map and the corresponding distance on the ground.
For example, on a \code{1:1000000} scale map, 1cm on the map equals 1km on the ground. Possible values are: \code{1, 3, 10, 20 or 60}, and
it's only for PROV or CCAA geographical granularity, \code{map_scale = 60} by default and \code{map_scale = NULL} for high detailed map.}

\item{verbose}{(boolean) show more information during the process}

\item{benchmark}{(boolean) used to measure the performance of the system, \code{benchmark = FALSE} by default.}

\item{det}{(int) \code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{type}{(string) what type of plot should be drawn, \code{type = "p"} (for points) by default. See \code{type} in \code{\link{plot}}}

\item{lang}{(string) language used to obtain information}
}
\description{
This function allows representing series data in maps or charts
}
\examples{
represent_series("IPC251521", resource = "maps")
represent_series(resource = "maps", help = TRUE)
represent_series("IPC251521", resource = "maps", map_scale = NULL)
represent_series("IPC206449", resource = "plot", nlast = 5)
represent_series("IPC206449", resource = "highcharts", nlast = 5)
}
