% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{update_series}
\alias{update_series}
\title{Update series}
\usage{
update_series(serie = NULL, benchmark = FALSE, page = 1, tip = "M",
  det = 2, lang = "ES")
}
\arguments{
\item{serie}{serie identificator}

\item{benchmark}{used to measure the performance of the system, \code{benchmark = FALSE} by default.}

\item{page}{\code{page = 1} to obtain data of an specific page.}

\item{tip}{\code{tip = M} to obtain the metadata (crossing variables-values) of the series.}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function allow update specific or all cache data
}
\examples{
update_series()
}
