% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{draw_serie}
\alias{draw_serie}
\title{Draw serie}
\usage{
draw_serie(serie, nult = 0, classification = NULL, verbose = FALSE,
  benchmark = FALSE)
}
\arguments{
\item{serie}{(string) serie id}

\item{nult}{(int) last \code{n} serie data, if \code{nult = 0} this value will be auto-calculated}

\item{classification}{(string) serie classification, if \code{classification = NULL} this value will be auto-detected}

\item{verbose}{(boolean) show more information during the process}
}
\description{
This function allows representing series data into a map
}
\examples{
draw_serie("IPC251541")
draw_serie("IPC251521")
draw_serie("UA42121")
}
