% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datos.R
\name{get_data_metadataoperation}
\alias{get_data_metadataoperation}
\title{Get data metadata-operation}
\usage{
get_data_metadataoperation(id, g1var = 0, g1val = 0, g2var = 0,
  g2val = 0, g3var = 0, p = 0, nult = 0, det = 0, tip = NA,
  ioe = FALSE, lang = "ES")
}
\arguments{
\item{id}{table identification}

\item{g1var}{variable (g1)}

\item{g1val}{valor (g1)}

\item{g2var}{variable (g2)}

\item{g2val}{valor (g2)}

\item{g3var}{variable (g3)}

\item{p}{periodicidad}

\item{nult}{last \code{n} values}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = AM} to obtain the metadata (crossing variables-values) of the series and a friendly output.}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with latest \code{n} series data by crossing metadata from an id or code
}
\examples{
get_data_metadataoperation("IPC", 115, 29, 3, 84, 762, p = 1, nult = 1)
}
