% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_natcode}
\alias{get_natcode}
\title{Get natcode}
\usage{
get_natcode(serie = NULL, variable_id = NULL, all = TRUE)
}
\arguments{
\item{serie}{(string) serie identificator}

\item{variable_id}{(int) variable identificator}

\item{all}{(bool) if \code{all = TRUE} you will get all natcodes}
}
\description{
This function allows get all natcodes or calculate a natcode from a serie and a geographical granularity
}
\examples{
get_natcode()
get_natcode("IPC251541", 115)
get_natcode("DPOP37286", 19)
get_natcode("IPC251522", 70)
}
