% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{get_natcode}
\alias{get_natcode}
\title{Get natcode}
\usage{
get_natcode(serie = NULL, all = TRUE, verbose = TRUE)
}
\arguments{
\item{serie}{(string) serie identificator}

\item{all}{(bool) if \code{all = TRUE} you will get all natcodes}

\item{verbose}{(boolean) show more information during the process}
}
\description{
This function allows get all natcodes or calculate a natcode from a serie and a geographical granularity
}
\examples{
get_natcode()
get_natcode("IPC251522")
get_natcode("IPC251541")
get_natcode("DPOP37286")
}
