% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operation}
\alias{get_operation}
\title{Get operation}
\usage{
get_operation(operation, ioe = FALSE, lang = "ES")
}
\arguments{
\item{operation}{operation identificator}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all available statistical operations from an id or code
}
\examples{
get_operation(operation = 25)
get_operation(operation = "IPC")
get_operation(operation = 30138, ioe = TRUE)
}
