% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{get_operation}
\alias{get_operation}
\title{Get operation}
\usage{
get_operation(code, ioe = FALSE, lang = "ES")
}
\arguments{
\item{code}{to find an operation}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all available statistical operations of an id or code
}
\details{
Numeric code (id)
Alphabetic code (IPC)
IOE code (Inventario de Operaciones Estadísticas)
}
\examples{
get_operation(25)
get_operation("IPC")
get_operation(30138, ioe = TRUE)
}
