% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{get_publications}
\alias{get_publications}
\title{Get publications}
\usage{
get_publications(det = 0, lang = "ES")
}
\arguments{
\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all available publications from an id or code
}
\examples{
get_publications()
}
