% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_serie}
\alias{get_serie}
\title{Get serie}
\usage{
get_serie(code, det = 0, tip = NA, lang = "ES", cache = FALSE,
  benchmark = FALSE)
}
\arguments{
\item{code}{serie identification}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = M} to obtain the metadata (crossing variables-values) of the series.}

\item{lang}{language used to obtain information}

\item{cache}{used to load data from local cache instead API, \code{cache = FALSE} by default.}

\item{benchmark}{used to measure the performance of the system, \code{benchmark = FALSE} by default.}
}
\description{
This function returns a data frame with a serie from an id or code
}
\examples{
get_serie("IPC206449")
get_serie("IPC206449", 2, "M")
}
