% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{get_series_metadataoperation}
\alias{get_series_metadataoperation}
\title{Get series metadata operation}
\usage{
get_series_metadataoperation(code, query = NULL, p = NULL, det = 0,
  tip = NA, ioe = FALSE, lang = "ES")
}
\arguments{
\item{code}{operation identification}

\item{query}{string separated by \code{AND} with syntax \code{variable = value} using natural language}

\item{p}{periodicity, \code{p = 1} by default}

\item{det}{\code{det = 2} to see two levels of depth, specifically to access the \code{PubFechaAct} object, \code{det = 0} by default}

\item{tip}{\code{tip = M} to obtain the metadata (crossing variables-values) of the series.}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all series by metadata crossing from an id or code
}
\details{
Numeric code \code{id}
Alphabetic code \code{IPC}
\code{IOE} code (Inventario de Operaciones Estadísticas)
Código identificativo de la operación (IOE30138 /IPC/ 25) y códigos identificativos de las variables y valores:
Provincias (FK_VARIABLE=115) = "Madrid" (FK_VALOR=29)  >  g1=115:29
Tipo de dato (FK_VARIABLE=3) = “Variación mensual”   (FK_VALOR=84) > g2=3:84
Grupos ECOICOP (FK_VARIABLE=762) = "Todos los grupos ECOICOP” (FK_VALOR=null) >  g3=762:
Serie mensual (FK_PERIODICIDAD=1) >  p=1  (Ver PUBLICACIONES_OPERACION)
}
\examples{
get_series_metadataoperation("IPC", query = "Provincias = Madrid AND Tipo de dato = Variación mensual AND Grupos ECOICOP = NULL")
}
