% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{get_var_values_operation}
\alias{get_var_values_operation}
\title{Get variable values operation}
\usage{
get_var_values_operation(operation, ioe = FALSE, lang = "ES")
}
\arguments{
\item{operation}{operation identifier}

\item{ioe}{\code{TRUE} if code is in format \code{IO30138}, and \code{FALSE} by default}

\item{lang}{language used to obtain information}
}
\description{
This function returns a data frame with all values and variables of an operation from an id or code
}
\examples{
get_var_values_operation(operation = 25)
get_var_values_operation(operation = "IPC")
get_var_values_operation(operation = 30138, ioe = TRUE)
}
