% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{update_cache}
\alias{update_cache}
\title{Update cache}
\usage{
update_cache(code = 0, n = 0, page = NA, pagination = TRUE,
  page_start = NA, page_end = NA, benchmark = TRUE)
}
\arguments{
\item{code}{operation identificator}

\item{n}{number of operation to update starting from first operation getted from \code{get_operations_all()} function.}

\item{page}{\code{page = 1} to obtain data of an specific page (to use this, \code{pagination = FALSE}).}

\item{pagination}{\code{TRUE} to obtain data page by page and \code{FALSE} by default.}

\item{page_start}{\code{page_start = 1} start page range to obtain data (to use this, \code{pagination = TRUE}).}

\item{page_end}{\code{page_end = 2} end page range to obtain data (to use this, \code{pagination = TRUE}).}

\item{benchmark}{used to measure the performance of the system, \code{benchmark = FALSE} by default.}
}
\description{
This function allow update specific or all cache data
}
\examples{
update_cache(code = 249)
update_cache(code = 249, page = 1)
update_cache(n = 3)
}
